requirements.txt
aniso8601==8.1.0
backcall==0.2.0
bcrypt==3.2.0
blessings==1.7
bpython==0.20.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
curtsies==0.3.4
decorator==4.4.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
openpyxl==3.0.6
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.11
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
traitlets==4.3.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
